+++ curr-app/src/components/CurrencySelect/select.module.scss
