--- curr-app/src/components/ui/ConverterInput/ConverterInput.tsx
@@ -1,5 +1,0 @@
-const ConverterInput = () => {
-  return <input type="text" defaultValue={1}></input>;
-};
-
-export default ConverterInput;
