+++ curr-app/src/pages/index.tsx
@@ -1,5 +1,5 @@
 import AllCurrencies from '@/components/AllCurrencies/AllCurrencies';
-import CurrencySelect from '@/components/ui/CurrencySelect/CurrencySelect';
+import CurrencySelect from '@/components/CurrencySelect/CurrencySelect';
 import style from '@/styles/home.module.scss';
 import Head from 'next/head';
 
@@ -8,7 +8,7 @@ export default function Home() {
     <>
       <Head>
         <title>Currency-app: Home</title>
-        <meta name="description" content="Generated by create next app" />
+        <meta name="description" content="Currency converter" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
